+++ packages/helpers/src/constants.ts
@@ -1,11 +1,8 @@
 export const CIRCOM_FIELD_MODULUS = 21888242871839275222246405745257275088548364400416034343698204186575808495617n;
-export const MAX_HEADER_PADDED_BYTES = 1024; // NOTE: this must be the same as the first arg in the email in main args circom
-export const MAX_BODY_PADDED_BYTES = 1536; // NOTE: this must be the same as the arg to sha the remainder number of bytes in the email in main args circom
+export const MAX_HEADER_PADDED_BYTES = 1024;  // Default value for max size to be used in circuit
+export const MAX_BODY_PADDED_BYTES = 1536;  // Default value for max size to be used in circuit
 
-// circom constants from main.circom / https://zkrepl.dev/?gist=30d21c7a7285b1b14f608325f172417b
-// template RSAGroupSigVerify(n, k, levels) {
-// component main { public [ modulus ] } = RSAVerify(121, 17);
-// component main { public [ root, payload1 ] } = RSAGroupSigVerify(121, 17, 30);
+
 export const CIRCOM_BIGINT_N = 121;
 export const CIRCOM_BIGINT_K = 17;
 export const CIRCOM_LEVELS = 30;
