+++ packages/helpers/src/sha-utils.ts
@@ -1,5 +1,5 @@
 import * as CryptoJS from 'crypto';
-import { assert, int64toBytes, int8toBytes, mergeUInt8Arrays } from "./binaryFormat";
+import { assert, int64toBytes, int8toBytes, mergeUInt8Arrays } from "./binary-format";
 import { Hash } from "./fast-sha256";
 
 export function findIndexInUint8Array(
